Split the RUSTFLAGS contract out of makefile_test_target
Closes #524. The file had grown to 732 lines against the repository's
400-line ceiling, which whitaker does not currently enforce on
crate-root test files. The RUSTFLAGS contract — the constants, the
RustflagsCase table, the expansion helpers, and the behavioural
suites — moves to a sibling module with its own module documentation;
the parent keeps the make test runner contract and the shared quoted
manifest flag, which the module imports.

The two twelve-case expansion suites shared their setup and split one
contract between them; they are now a single parameterized test that
expands each recipe twice — with and without inherited flags — and
asserts both halves, so the per-case failure message still names the
target while the Makefile is read once per case rather than twice.
Both files now sit under the ceiling (320 and 392 lines).
